Add setup/teardown

+++ libkbfs/journal_md_ops_test.go
@@ -15,22 +15,25 @@ import (
 	"golang.org/x/net/context"
 )
 
-// TODO: Clean up the test below.
+func setupJournalMDOpsTest(t *testing.T) (
+	tempdir string, config Config, oldMDOps MDOps, jServer *JournalServer) {
+	config = MakeTestConfigOrBust(t, "test_user")
 
-func TestJournalMDOpsBasics(t *testing.T) {
-	// setup
 	tempdir, err := ioutil.TempDir(os.TempDir(), "journal_md_ops")
 	require.NoError(t, err)
+	// Clean up the tempdir if anything in the setup fails/panics.
 	defer func() {
-		err := os.RemoveAll(tempdir)
-		require.NoError(t, err)
+		if r := recover(); r != nil {
+			CheckConfigAndShutdown(t, config)
+			err := os.RemoveAll(tempdir)
+			if err != nil {
+				t.Errorf(err.Error())
+			}
+		}
 	}()
 
-	config := MakeTestConfigOrBust(t, "test_user")
-	defer CheckConfigAndShutdown(t, config)
-
 	log := config.MakeLogger("")
-	jServer := makeJournalServer(
+	jServer = makeJournalServer(
 		config, log, tempdir, config.BlockCache(),
 		config.BlockServer(), config.MDOps())
 
@@ -41,11 +44,25 @@ func TestJournalMDOpsBasics(t *testing.T) {
 	config.SetBlockCache(jServer.blockCache())
 	config.SetBlockServer(jServer.blockServer())
 
-	oldMDOps := config.MDOps()
+	oldMDOps = config.MDOps()
 	config.SetMDOps(jServer.mdOps())
 
-	mdOps := config.MDOps()
+	return tempdir, config, oldMDOps, jServer
+}
 
+func teardownJournalMDOpsTest(t *testing.T, tempdir string, config Config) {
+	CheckConfigAndShutdown(t, config)
+	err := os.RemoveAll(tempdir)
+	require.NoError(t, err)
+}
+
+// TODO: Clean up the test below.
+
+func TestJournalMDOpsBasics(t *testing.T) {
+	tempdir, config, oldMDOps, jServer := setupJournalMDOpsTest(t)
+	defer teardownJournalMDOpsTest(t, tempdir, config)
+
+	ctx := context.Background()
 	_, uid, err := config.KBPKI().GetCurrentUserInfo(ctx)
 	require.NoError(t, err)
 
@@ -55,6 +72,8 @@ func TestJournalMDOpsBasics(t *testing.T) {
 
 	h, err := MakeTlfHandle(ctx, bh, config.KBPKI())
 	require.NoError(t, err)
+
+	mdOps := jServer.mdOps()
 
 	id, irmd, err := mdOps.GetForHandle(ctx, h, Merged)
 	require.NoError(t, err)
